requirements.txt
# Core
Flask==2.3.3
flask-cors==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Database
supabase==2.0.0
postgrest==0.13.0

# Auth & Security
PyJWT==2.8.0
cryptography==41.0.7

# Redis & Cache
redis==5.0.1
hiredis==2.2.3

# Async & Tasks
celery==5.3.4
kombu==5.3.4
flower==2.0.1

# Utils
requests==2.31.0
python-dateutil==2.8.2

# Telegram
python-telegram-bot==20.6

# TON (якщо будете інтегрувати)
# tonsdk==1.0.13
# pytonlib==0.0.60

# Monitoring (опціонально)
sentry-sdk[flask]==1.38.0